requirements.txt
numpy==2.2.3
pillow==11.1.0
PyMuPDF==1.25.4
pyserial==3.5
